Split config_loader
