+++ src/mantis/jira/jira_issues.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING, Any
 from pydantic import BaseModel
 
 from drafts import Draft
+from mantis.jira.config_loader.meta_model_factories import CreatemetaModelFactory, EditmetaModelFactory
 from mantis.jira.issue_field import IssueField
-from mantis.jira.config_loader.config_loader import CreatemetaModelFactory, EditmetaModelFactory
 
 if TYPE_CHECKING:
     from .jira_client import JiraClient
